backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
flask-sqlalchemy==2.5.1
tensorflow==2.10.0
librosa==0.9.1
numpy==1.23.1
Pillow==9.0.1
PyJWT==2.3.0
gunicorn==20.1.0
werkzeug==2.0.2
scipy==1.8.1
soundfile==0.10.3.post1
librosa==0.9.1
numpy==1.23.1
Pillow==9.0.1
PyJWT==2.3.0
gunicorn==20.1.0
werkzeug==2.0.2